🚚: Reorganization of some pages
