+++ frontend/app_student/lib/class_groups/views/class_group.dart
