+++ frontend/app_student/lib/login/views/widgets/form/button_submit.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
-import '../../cubit/login_cubit.dart';
+import '../../../cubit/login_cubit.dart';
 
 class SubmitButton extends StatelessWidget {
   final TextEditingController ineController;
